+++ RVM/log/LogEntry.java
@@ -1,4 +1,4 @@
-package rvm;
+package log;
 
 import java.util.HashMap;
 import java.util.List;
