+++ test/TestLogMonitor.java
@@ -1,14 +1,14 @@
-import junit.framework.Assert;
+import analysis.LogMonitor;
 import org.junit.Test;
 
 /**
  * Created by xiaohe on 14-11-24.
  */
-public class TestLogEntryExtractor {
+public class TestLogMonitor {
     @Test
     public void testLogEntryMain(){
         String[] args=new String[]{};
-        LogEntryExtractor.main(args);
+        LogMonitor.main(args);
 
         //after refactoring, assert whether the log file has been fully read
         //by testing whether hasNext() returns false;
